Moved ViewModels into separate folder
