+++ Sources/Recycling/ListViews/Generic/Cells/Demo/CheckboxCellDemoView.swift
@@ -4,7 +4,7 @@
 
 import FinniversKit
 
-private struct ViewModel: CheckboxTableViewCellViewModel {
+private struct ViewModel: SelectableTableViewCellViewModel {
     let title: String
     let subtitle: String? = nil
     let hasChevron: Bool = false
