+++ Sources/Recycling/ListViews/Generic/ViewModels/BasicTableViewCellViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import Foundation
+
+public protocol BasicTableViewCellViewModel {
+    var title: String { get }
+    var subtitle: String? { get }
+    var hasChevron: Bool { get }
+}
