+++ Sources/Recycling/ListViews/Generic/ViewModels/SelectableTableViewCellViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import Foundation
+
+public protocol SelectableTableViewCellViewModel: BasicTableViewCellViewModel {
+    var isSelected: Bool { get }
+}
